Remove user-switching code
The existing setup is incompatible with Podman's rootless mode. When the
container is _already_ running as a non-root user, then there's no need
to create another user inside the container, and no need to re-assign
ownership of all the files on the mounted volumes. Instead, it's safe to
assume that the container user is already the desired user, and that it
already owns all the files it needs.

The only role of the entrypoint script was to set up the user and group,
so once those commands were removed, all that remained was the single
node command. I've deleted the entrypoint script and am directly running
node as the entrypoint instead.

I have no idea what effect this has on Docker when it starts each
container as root.