Avoid creating wrong config directory
Paths for assets, logs, and database are defaulted to the home dot-file
directory only if corresponding environment variables aren't set.
Regardless, the dot-file directory is checked for existence and created
even when all the environment variables point somewhere else.

This change disregards the specfic home directory and instead checks
whether the parents of _any_ of the config direvtories need to be
created.

This makes particular difference when the program is running as a user
who doesn't have a home directory.