Remove hostname specification on listen
When a hostname is provided here, I think it ends up resolving that
hostname to its IP address, and then the server listens _only_ on that
IP address. That means it doesn't listen on localhost. This is a
particular problem when there's a Tailscale sidecar trying to redirect
connections to 127.0.0.1:7418 and it gets "connection refused" because
there's nothing listening there. Omitting the hostname causes the server
to listen on all addresses.